Refactoring - remove SchemaBuilderHelpers and use ResultHelpersin SchemaExectionTests (#6)
